requirements.txt
Django==1.11.2
django-filter==1.0.4
djangorestframework==3.6.3
psycopg2==2.7.1
pytz==2017.2
